Moved all shared lambda types into a single location so it's easier later to refactor it later to some better structure Added improved error handling to the portfolio page
