--- lambdas/src/common/domain.ts
@@ -1,21 +1,0 @@
-export enum NetworkNames {
-  ethereumMainnet = 'ethereum',
-  ethereumGoerli = 'ethereum_goerli',
-  arbitrumMainnet = 'arbitrum',
-  arbitrumGoerli = 'arbitrum_goerli',
-  polygonMainnet = 'polygon',
-  polygonMumbai = 'polygon_mumbai',
-  optimismMainnet = 'optimism',
-  optimismGoerli = 'optimism_goerli',
-  baseMainnet = 'base',
-  baseGoerli = 'base_goerli',
-}
-
-export enum LendingProtocol {
-  AaveV2 = 'aavev2',
-  AaveV3 = 'aavev3',
-  Ajna = 'ajna',
-  Maker = 'maker',
-  MorphoBlue = 'morphoblue',
-  SparkV3 = 'sparkv3',
-}
