--- lambdas/src/portfolio-overview/types.ts
@@ -1,9 +1,0 @@
-export type PortfolioOverviewResponse = {
-  walletBalanceUsdValue: number
-  suppliedUsdValue: number
-  suppliedPercentageChange: number
-  borrowedUsdValue: number
-  borrowedPercentageChange: number
-  totalUsdValue: number
-  totalPercentageChange: number
-}
