+++ lambdas/src/portfolio-positions/index.ts
@@ -1,9 +1,9 @@
 /* eslint-disable no-relative-import-paths/no-relative-import-paths */
 import type { APIGatewayProxyEventV2, APIGatewayProxyResultV2 } from 'aws-lambda'
 
-import { getDefaultErrorMessage } from '../common/helpers'
-import { ResponseBadRequest, ResponseOk } from '../common/responses'
-import { getAddressFromRequest } from '../common/validators'
+import { getDefaultErrorMessage } from '../shared/helpers'
+import { ResponseBadRequest, ResponseOk } from '../shared/responses'
+import { getAddressFromRequest } from '../shared/validators'
 import { mockPositionsResponse } from './mock'
 
 export const handler = async (event: APIGatewayProxyEventV2): Promise<APIGatewayProxyResultV2> => {
