+++ lambdas/src/shared/constants.ts
