+++ lambdas/src/shared/errors.ts
