+++ lambdas/src/shared/guards.ts
