+++ lambdas/src/shared/helper-types.ts
