+++ lambdas/src/shared/helpers.ts
